feat: enhance footer navigation and add header-only menu items
- Updated Footer tests to ensure header-only menu items are not rendered.
- Introduced  to exclude specific links from the footer.
- Added new fields to GossipField for charter, committee, and governance.
- Updated navigation data to reflect changes in menu items.
- Incremented organization and repository stats for accuracy.
- Added end-to-end tests for TSC profile dialog focus management.
- Implemented TSC Committee component with member profiles and dialog functionality.
- Created TSC Hero Network component for animated background effect.
