+++ src/app/__tests__/component-classes.test.ts
@@ -11,7 +11,7 @@ import { describe, expect, it } from "vitest";
  * This closes that gap. It deliberately covers only the prefixes that follow
  * the convention, so it does not object to ordinary Tailwind utilities.
  */
-const PREFIXES = ["menu-", "site-header", "blog-", "issues-"];
+const PREFIXES = ["menu-", "site-header", "blog-", "issues-", "tsc-"];
 const SRC = path.resolve(__dirname, "../..");
 const CSS = path.resolve(__dirname, "../globals.css");
