+++ src/data/organization_stats.json
@@ -1,1 +1,1 @@
-{ "publicRepositories": 42, "totalStars": 3030 }
+{ "publicRepositories": 42, "totalStars": 3033 }
