+++ src/data/repository_stats.json
@@ -2,7 +2,7 @@
   "hiero-consensus-node": { "stars": 404 },
   "hiero-local-node": { "stars": 262 },
   "hiero-mirror-node": { "stars": 185 },
-  "hiero-improvement-proposals": { "stars": 180 },
+  "hiero-improvement-proposals": { "stars": 181 },
   "hiero-sdk-js": { "stars": 328 },
   "hiero-sdk-java": { "stars": 262 },
   "hiero-json-rpc-relay": { "stars": 92 },
@@ -13,7 +13,7 @@
   "solo": { "stars": 43 },
   "hiero-block-node": { "stars": 38 },
   "hiero-sdk-tck": { "stars": 24 },
-  "hiero-sdk-cpp": { "stars": 41 },
+  "hiero-sdk-cpp": { "stars": 42 },
   "governance": { "stars": 14 },
   "hiero-sdk-python": { "stars": 63 },
   "hiero-sdk-swift": { "stars": 37 },
